requirements.txt
pillow
django
django-cleanup
django-ninja